# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LoanInterest:
    properties:
      duration mounths:
        type: integer
      interest rate:
        type: number
    type: object
...
  dto.Paging:
    properties:
      Page:
        type: integer
      size:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  dto.ResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  dto.ResponseMessage:
    properties:
      message:
        type: string
    type: object
  dto.ResponsePaging:
    properties:
      data: {}
      message:
        type: string
      paging:
        $ref: '#/definitions/dto.Paging'
    type: object
...
  model.LoanInterest:
    properties:
      duration mounths:
        type: integer
      id:
        type: string
      interest rate:
        type: number
    type: object
info:
  contact: {}
  title: Laundry Apps
  version: "1.0"
paths:
...
  /loaninterest:
    post:
      consumes:
      - application/json
      description: create new data loan interest
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data loan interest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoanInterest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: new
      tags:
      - loan interest
    put:
      consumes:
      - application/json
      description: create update data loan interest
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data loan interest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoanInterest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: update
      tags:
      - loan interest
  /loaninterest/{id}:
    delete:
      consumes:
      - application/json
      description: get all data loan interest
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponsePaging'
      summary: get all
      tags:
      - loan interest
  /loaninterest/list/{page}/{size}:
    get:
      consumes:
      - application/json
      description: get all data loan interest
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page of pagination
        in: path
        name: page
        required: true
        type: integer
      - description: size of pagination
        in: path
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponsePaging'
      summary: get all
      tags:
      - loan interest
  /saldo:
    get:
      consumes:
      - application/json
      description: get by id data loan
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: get by id installenment
      tags:
      - saldo
  /saldo/{id}:
    get:
      consumes:
      - application/json
      description: get by id data loan
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: get by id installenment
      tags:
      - saldo
  /saldo/list/{page}/{size}:
    get:
      consumes:
      - application/json
      description: get all data saldo
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page of pagination
        in: path
        name: page
        required: true
        type: integer
      - description: size of pagination
        in: path
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponsePaging'
      summary: get all
      tags:
      - saldo
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"